+++ examples/multi_objective/batch_simple.py
@@ -22,7 +22,7 @@ def callback(study, trial):
 
 if __name__ == "__main__":
     study = optuna.multi_objective.create_study(["minimize", "minimize"])
-    bstudy = optuna.BatchMultiObjectiveStudy(study, batch_size=4)
+    bstudy = optuna.multi_objective.BatchMultiObjectiveStudy(study, batch_size=4)
     bstudy.batch_optimize(objective, n_batches=10, callbacks=[callback])
 
     optuna.multi_objective.visualization.plot_pareto_front(study).show()
