+++ optuna/_batch_study.py
@@ -2,12 +2,11 @@ import copy
 import gc
 import math
 from typing import Callable
-from typing import List  # NOQA
-from typing import Optional  # NOQA
-from typing import Sequence  # NOQA
-from typing import Tuple  # NOQA
-from typing import Type  # NOQA
-from typing import Union  # NOQA
+from typing import List
+from typing import Optional
+from typing import Tuple
+from typing import Type
+from typing import Union
 
 import numpy as np
 
@@ -15,22 +14,12 @@ import optuna
 from optuna import exceptions
 from optuna import logging
 from optuna.multi_objective.study import MultiObjectiveStudy
-from optuna.multi_objective.trial import FrozenMultiObjectiveTrial
 from optuna import trial as trial_module
-from optuna.multi_objective.trial import BatchMultiObjectiveTrial
 from optuna.trial import BatchTrial
 
 BatchObjectiveFuncType = Callable[[BatchTrial], np.ndarray]
-BatchMultiObjectiveFuncType = Callable[[BatchMultiObjectiveTrial], Sequence[np.ndarray]]
 CallbackFuncType = Optional[
     List[Callable[["optuna.study.Study", "optuna.trial.FrozenTrial"], None]]
-]
-MultiObjectiveCallbackFuncType = Callable[
-    [
-        "optuna.multi_objective.study.MultiObjectiveStudy",
-        "optuna.multi_objective.trial.FrozenMultiObjectiveTrial",
-    ],
-    None,
 ]
 ObjectiveFuncType = Callable[[trial_module.Trial], float]
 
@@ -203,58 +192,3 @@ class BatchStudy(optuna.study.Study):
             self._storage.set_trial_state(trial_id, trial_module.TrialState.COMPLETE)
             self._log_completed_trial(trial, result)
         return trials
-
-
-class BatchMultiObjectiveStudy(MultiObjectiveStudy):
-    def __init__(self, study: MultiObjectiveStudy, batch_size: int) -> None:
-        self._sampler = study.sampler
-        super().__init__(BatchStudy(study, batch_size))
-
-    @property
-    def sampler(self) -> "optuna.multi_objective.samplers.BaseMultiObjectiveSampler":
-        return self._sampler
-
-    def batch_optimize(
-        self,
-        objective: BatchMultiObjectiveFuncType,
-        timeout: Optional[int] = None,
-        n_batches: Optional[int] = None,
-        n_jobs: int = 1,
-        catch: Tuple[Type[Exception], ...] = (),
-        callbacks: Optional[List[MultiObjectiveCallbackFuncType]] = None,
-        gc_after_trial: bool = True,
-        show_progress_bar: bool = False,
-    ) -> None:
-        def mo_objective(trial: BatchTrial) -> np.ndarray:
-            trials = [
-                optuna.multi_objective.trial.MultiObjectiveTrial(trial) for trial in trial._trials
-            ]
-            mo_trial = BatchMultiObjectiveTrial(trials)
-            values = objective(mo_trial)
-            mo_trial._report_complete_values(values)
-            return np.zeros(len(trials))  # Dummy value.
-
-        # Wraps a multi-objective callback so that we can pass it to the `Study.optimize` method.
-        def wrap_mo_callback(
-            callback: MultiObjectiveCallbackFuncType,
-        ) -> Callable[["optuna.Study", "optuna.trial.FrozenTrial"], None]:
-            return lambda study, trial: callback(
-                MultiObjectiveStudy(study), FrozenMultiObjectiveTrial(self.n_objectives, trial),
-            )
-
-        if callbacks is None:
-            wrapped_callbacks = None
-        else:
-            wrapped_callbacks = [wrap_mo_callback(callback) for callback in callbacks]
-
-        assert isinstance(self._study, BatchStudy)
-        self._study.batch_optimize(
-            mo_objective,
-            timeout=timeout,
-            n_batches=n_batches,
-            n_jobs=n_jobs,
-            catch=catch,
-            callbacks=wrapped_callbacks,
-            gc_after_trial=gc_after_trial,
-            show_progress_bar=show_progress_bar,
-        )
