+++ optuna/multi_objective/__init__.py
@@ -4,5 +4,6 @@ from optuna.multi_objective import study  # NOQA
 from optuna.multi_objective import trial  # NOQA
 from optuna.multi_objective import visualization  # NOQA
 
+from optuna.multi_objective._batch_study import BatchMultiObjectiveStudy  # NOQA
 from optuna.multi_objective.study import create_study  # NOQA
 from optuna.multi_objective.study import load_study  # NOQA
